refact: move classes into engine package
