+++ pathetic-bukkit/src/main/java/de/metaphoriker/pathetic/bukkit/mapper/BukkitMapper.java
@@ -14,7 +14,7 @@ import de.metaphoriker.pathetic.api.wrapper.PathBlock;
 import de.metaphoriker.pathetic.api.wrapper.PathEnvironment;
 import de.metaphoriker.pathetic.api.wrapper.PathPosition;
 import de.metaphoriker.pathetic.api.wrapper.PathVector;
-import de.metaphoriker.pathetic.util.ErrorLogger;
+import de.metaphoriker.pathetic.engine.util.ErrorLogger;
 
 @UtilityClass
 public class BukkitMapper {
