+++ pathetic-bukkit/src/main/java/de/metaphoriker/pathetic/bukkit/provider/FailingBlockProvider.java
@@ -8,7 +8,7 @@ import de.metaphoriker.pathetic.bukkit.provider.world.WorldDomain;
 import de.metaphoriker.pathetic.bukkit.util.BukkitVersionUtil;
 import de.metaphoriker.pathetic.bukkit.util.ChunkUtil;
 import de.metaphoriker.pathetic.resolver.ChunkDataProviderResolver;
-import de.metaphoriker.pathetic.util.ErrorLogger;
+import de.metaphoriker.pathetic.engine.util.ErrorLogger;
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
