+++ pathetic-engine/src/main/java/de/metaphoriker/pathetic/engine/util/ExpiringHashMap.java
@@ -1,4 +1,4 @@
-package de.metaphoriker.pathetic.util;
+package de.metaphoriker.pathetic.engine.util;
 
 import java.util.concurrent.ConcurrentHashMap;
 import lombok.Getter;
